fix: move discord related function into separate file
